refactor: reorganize Moderation Feature with Feature-Sliced Design structure
- Reorganize Features/Moderation/ with numbered file structure:
  - 01-ModerationFeature.cs: DI registration
  - 02-ModerationFacade.cs: feature entry point (scenario orchestration)
  - 03-Policies.cs: rules/decisions (core business logic)
  - 04-Services.cs: thin adapters to Notification/UserBan/etc.
  - 05-Contracts.cs: DTOs/interfaces for the feature
- Consolidate interfaces into 05-Contracts.cs
- Remove redundant interface files
- Maintain zero-diff functionality while improving organization
- Follow Feature-Sliced Design principles for better maintainability

This structure provides clear separation of concerns and makes the feature
more organized and easier to navigate.
